+++ src/emittery/src/init.lua
@@ -1,3 +1,4 @@
+--!nonstrict
 -- ROBLOX upstream: https://github.com/sindresorhus/emittery/blob/v0.11.0/index.js
 -- ROBLOX upstream types: https://github.com/sindresorhus/emittery/blob/v0.11.0/index.d.ts
 --[[