+++ src/roblox-shared/src/Writeable.lua
@@ -22,7 +22,7 @@ export type Writeable = {
 local Writeable = {}
 Writeable.__index = Writeable
 
-function Writeable.new(options: { write: (data: string) -> () }): Writeable
+function Writeable.new(options: { write: (data: string) -> () }?): Writeable
 	local self = setmetatable({}, Writeable)
 	self._writeFn = if options ~= nil and typeof(options.write) == "function" then options.write else print
 	self.isTTY = false